projects/project-201-distributed-training/requirements.txt
# Core ML Frameworks
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Distributed Training
ray[train]>=2.6.0
ray[tune]>=2.6.0

# MLOps
mlflow>=2.5.0
optuna>=3.2.0

# Data Processing
pandas>=2.0.0
pillow>=9.5.0
h5py>=3.8.0

# Monitoring
prometheus-client>=0.17.0
py-cpuinfo>=9.0.0
psutil>=5.9.5
pynvml>=11.5.0

# Kubernetes
kubernetes>=27.2.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.4.0
tqdm>=4.65.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Linting and Formatting
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
isort>=5.12.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0
